Refactor codebase into modular package structure with separation of concerns
Split monolithic main.py (300+ lines) into logical modules:
- certificate/: ASN.1 parsing and SSL certificate retrieval
- processing/: Domain processing, SAN extraction, and concurrent execution
- output/: Result formatting and file I/O
- cli/: Typer-based command-line interface

Each module includes proper __init__.py with explicit exports, comprehensive
docstrings, and maintains strict type safety. The refactoring eliminates
circular dependencies while preserving all existing functionality and CLI
behavior. Updated CLAUDE.md documentation to reflect BasedPyright usage.

All quality checks pass:
- BasedPyright strict mode: 0 errors
- Ruff linting: All checks passed
- CLI functionality: Verified working
