+++ src/gsan/output/__init__.py
@@ -1,0 +1,9 @@
+"""Output formatting and result presentation.
+
+This module handles formatting and outputting results in various formats
+(text, JSON) to console or file.
+"""
+
+from gsan.output.formatter import output_results
+
+__all__ = ["output_results"]
